refactor circuit_templates module
