--- cirkit/templates/circuit_templates/__init__.py
@@ -1,2 +1,0 @@
-from .data import image_data as image_data
-from .utils import Parameterization as Parameterization
