+++ cirkit/templates/utils.py
